+++ chemprop/v2/featurizers/mixins.py
@@ -1,15 +1,11 @@
-from abc import ABC, abstractmethod
-from typing import Optional, Sequence, Union
-
-import numpy as np
-from rdkit import Chem
+from typing import Optional
 
 from chemprop.v2.featurizers.molgraph import MolGraph
 from chemprop.v2.featurizers.atom import AtomFeaturizer, AtomFeaturizerBase
 from chemprop.v2.featurizers.bond import BondFeaturizer, BondFeaturizerBase
 
 
-class MolGraphFeaturizerMixin(ABC):
+class MolGraphFeaturizerMixin:
     def __init__(
         self,
         atom_featurizer: Optional[AtomFeaturizerBase] = None,
@@ -22,8 +18,7 @@ class MolGraphFeaturizerMixin(ABC):
         self.bond_fdim = len(self.bond_featurizer)
         self.bond_messages = bond_messages
 
-        
-    def __call__(self, *args, **kwargs):
+    def __call__(self, *args, **kwargs) -> MolGraph:
         return self.featurize(*args, **kwargs)
 
     @property
@@ -31,46 +26,3 @@ class MolGraphFeaturizerMixin(ABC):
         """the feature dimension of the atoms and bonds, respectively, of `MolGraph`s generated by
         this featurizer"""
         return self.atom_fdim, self.bond_fdim
-
-    @abstractmethod
-    def featurize(
-        self,
-        mol_or_reaction: Union[Chem.Mol, tuple[Chem.Mol, Chem.Mol]],
-        atom_features_extra: Optional[np.ndarray] = None,
-        bond_features_extra: Optional[np.ndarray] = None,
-    ) -> MolGraph:
-        pass
-
-
-class MultiHotFeaturizerMixin(ABC):
-    """A `MultiHotFeaturizer` calculates feature vectors of arbitrary objects by concatenating
-    multiple one-hot feature vectors"""
-
-    def __call__(self, x) -> np.ndarray:
-        return self.featurize(x)
-
-    @abstractmethod
-    def __len__(self) -> int:
-        """the length of a feature vector from this featurizer"""
-
-    @property
-    @abstractmethod
-    def subfeatures(self) -> Sequence[tuple[str, slice]]:
-        """a list of tuples containing the subfeature name and its respective slice in the output 
-        feature vectors"""
-
-    @abstractmethod
-    def featurize(self, x) -> np.ndarray:
-        """calculate the feature vector of x"""
-
-    @property
-    def num_subfeatures(self) -> int:
-        return len(self.subfeatures)
-
-    @staticmethod
-    def one_hot_index(x, xs: Sequence) -> tuple[int, int]:
-        """return the index of a one hot encoding of `x` given choices `xs` and the length of the
-        uncompressed encoding"""
-        n = len(xs)
-        
-        return xs.index(x) if x in xs else n, n + 1
